chore: rename share folder
