+++ editors/shared/components/Layout.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import type React from 'react';
 import ContentCard from './content-card.js';
 
 interface LayoutProps {
