+++ editors/shared/components/content-card.tsx
